requirements.txt
torch>=2.1.0
torchvision>=0.16.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
numpy>=1.26.0
Pillow>=10.1.0
python-dotenv>=1.0.0
pytest>=7.4.0
httpx>=0.26.0
